benchmark(overhead): tests + move overhead logic to test file
