+++ codecov.yml
@@ -1,6 +1,7 @@
 ignore:
   - "env"
   - "mocks"
+  - "primitives/benchmarking"
 
 coverage:
   precision: 2
